+++ src/components/Timesheet.tsx
@@ -1,12 +1,16 @@
-import React from 'react';
+import React, { useContext } from 'react';
+import { DashboardContext } from '@pages/Dashboard';
 import { StyledTimesheet, StyledRecord } from '@styles/components/Timesheet.styled';
 
 type TimesheetProps = {
   children: React.ReactNode;
 };
 
-const Record = (): JSX.Element =>
-  <StyledRecord>Time Record</StyledRecord>;
+const Record = (): JSX.Element => {
+  const context = useContext(DashboardContext);
+  return <StyledRecord>{ context }</StyledRecord>;
+};
+
 
 const Timesheet = (props: TimesheetProps): JSX.Element =>
   <StyledTimesheet>{ props.children }</StyledTimesheet>;